# internal/dto/rbac.go
package dto

// AddRoleDTO 添加角色DTO
type AddRoleDTO struct {
	Name    string `json:"name" validate:"required,min=1,max=50" label:"角色名称"`
	Remark  string `json:"remark" validate:"omitempty,max=200" label:"角色备注"`
	Status  int    `json:"status" validate:"oneof=0 1" label:"状态"` // 0-禁用 1-启用
	MenuIDs []int  `json:"menu_ids" validate:"omitempty" label:"菜单ID列表"`
}

type EditRoleDTO struct {
	ID      int     `uri:"id" validate:"required" label:"角色ID"`
	Name    *string `json:"name" validate:"omitempty,min=1,max=50" label:"角色名称"`
	Remark  *string `json:"remark" validate:"omitempty,max=200" label:"角色备注"`
	Status  *int    `json:"status" validate:"omitempty,oneof=0 1" label:"状态"` // 0-禁用 1-启用
	MenuIDs []int   `json:"menu_ids" validate:"omitempty" label:"菜单ID列表"`
}

// AssignRolesDTO 分配角色给用户DTO
type AssignRolesDTO struct {
	UserID  int   `json:"user_id" validate:"required" label:"用户ID"`
	RoleIDs []int `json:"role_ids" validate:"required,min=1" label:"角色ID列表"`
}

// AssignMenusDTO 分配菜单给角色DTO
type AssignMenusDTO struct {
	RoleID  int   `json:"role_id" validate:"required" label:"角色ID"`
	MenuIDs []int `json:"menu_ids" validate:"required,min=1" label:"菜单ID列表"`
}

// AssignPermissionsToMenuDTO 分配权限给菜单DTO（菜单-权限键多对多，同一权限键可绑多个菜单）
type AssignPermissionsToMenuDTO struct {
	MenuID   int      `json:"menu_id" validate:"required" label:"菜单ID"`
	PermKeys []string `json:"perm_keys" validate:"required" label:"权限键列表(path#method)"`
}

// GetRoleDTO 获取角色DTO
type GetRoleDTO struct {
	ID int `uri:"id" validate:"required" label:"角色ID"`
}

// GetUserRolesDTO 获取用户角色DTO
type GetUserRolesDTO struct {
	UserID int `uri:"id" validate:"required" label:"用户ID"`
}

// GetRolePermissionsDTO 获取角色权限DTO
type GetRolePermissionsDTO struct {
	RoleID int `uri:"id" validate:"required" label:"角色ID"`
}

// DeletePermissionDTO 解绑菜单权限键DTO
type DeletePermissionDTO struct {
	MenuID  int    `json:"menu_id" validate:"required" label:"菜单ID"`
	PermKey string `json:"perm_key" validate:"required" label:"权限键(path#method)"`
}