# internal/swagger/swagger.yaml
basePath: /
definitions:
...
  dto.AssignPermissionsToMenuDTO:
    properties:
      menu_id:
        type: integer
      perm_keys:
        items:
          type: string
        type: array
    required:
    - menu_id
    - perm_keys
    type: object
...
  dto.MenuDTO:
    properties:
      component:
        type: string
      meta:
        $ref: '#/definitions/models.MenuMeta'
      name:
        type: string
      path:
        type: string
      pid:
        type: integer
      status:
        description: 0-禁用 1-启用
        enum:
        - 0
        - 1
        type: integer
      type:
        enum:
        - catalog
        - menu
        - button
        - embedded
        - link
        type: string
    required:
    - meta
    - name
    - status
    - type
    type: object
...
  models.Menu:
    properties:
      children:
        items:
          $ref: '#/definitions/models.Menu'
        type: array
      component:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      meta:
        $ref: '#/definitions/models.MenuMeta'
      name:
        type: string
      path:
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.MenuPermission'
        type: array
      pid:
        type: integer
      redirect:
        type: string
      status:
        type: integer
      type:
        type: string
      updated_at:
        type: string
    type: object
  models.MenuMeta:
    properties:
      affixTab:
        type: boolean
      authority:
        items:
          type: string
        type: array
      badge:
        type: string
      badgeType:
        type: string
      badgeVariants:
        type: string
      hideChildrenInMenu:
        type: boolean
      hideInBreadcrumb:
        type: boolean
      hideInMenu:
        type: boolean
      hideInTab:
        type: boolean
      icon:
        type: string
      iframeSrc:
        type: string
      keepAlive:
        type: boolean
      link:
        type: string
      menuVisibleWithForbidden:
        type: boolean
      sort:
        type: integer
      title:
        type: string
    type: object
  models.MenuPermission:
    properties:
      description:
        type: string
      menu_id:
        type: integer
      perm_key:
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      request_id:
        type: string
    type: object
info:
  contact: {}
  description: webgos 企业资源计划系统 API 文档
  title: webgos API
  version: "1.0"
paths:
...
  /api/menu:
    post:
      consumes:
      - application/json
      description: 创建新菜单
      parameters:
      - description: 菜单参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.MenuDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建菜单
      tags:
      - 菜单管理
  /api/menu/:id:
    delete:
      consumes:
      - application/json
      description: 删除菜单
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除菜单
      tags:
      - 菜单管理
    get:
      consumes:
      - application/json
      description: 获取菜单详情
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 菜单详情
      tags:
      - 菜单管理
  /api/menu/{id}:
    put:
      consumes:
      - application/json
      description: 编辑菜单信息
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 菜单参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.MenuDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 编辑菜单
      tags:
      - 菜单管理
  /api/menu/list:
    get:
      consumes:
      - application/json
      description: 获取所有菜单列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取菜单列表
      tags:
      - 菜单管理
  /api/menu/name_exists:
    get:
      consumes:
      - application/json
      description: 检查菜单名称是否存在
      parameters:
      - description: 菜单名称
        in: query
        name: name
        required: true
        type: string
      - description: 菜单ID（编辑时用于排除自身）
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 检查菜单名称是否存在
      tags:
      - 菜单管理
  /api/menu/path_exists:
    get:
      consumes:
      - application/json
      description: 检查菜单路径是否存在
      parameters:
      - description: 菜单路径
        in: query
        name: path
        required: true
        type: string
      - description: 菜单ID（编辑时用于排除自身）
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 检查菜单路径是否存在
      tags:
      - 菜单管理
  /api/menu/permissions:
    post:
      consumes:
      - application/json
      description: 将权限点集合绑定到指定菜单
      parameters:
      - description: 绑定信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AssignPermissionsToMenuDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 绑定菜单权限
      tags:
      - 菜单管理
  /api/menu/permissions/{id}:
    get:
      consumes:
      - application/json
      description: 获取指定菜单所绑定的所有权限点（多对多）
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取菜单权限
      tags:
      - 菜单管理
  /api/menu/tree:
    get:
      consumes:
      - application/json
      description: 获取菜单树结构
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取菜单树
      tags:
      - 菜单管理
  /api/menu/user_menus:
    get:
      consumes:
      - application/json
      description: 根据当前登录用户的ID，获取其拥有的角色所关联的菜单权限树
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Menu'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取当前用户的菜单树
      tags:
      - 菜单管理
...
securityDefinitions:
  BearerAuth:
    description: JWT认证方式，值为"Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"